Refatora tipos e concorrência; melhora importação de contas
- Altera campos Saldo e Valor de decimal para double.
- Garante thread-safety na soma do saldo usando Interlocked.
- Usa double.Parse em addOperacaoLine.
- Substitui List<Conta> por Dictionary<int, Conta> em ContasManager.
- Comenta atualização visual do DataGrid.
- Adiciona exibição de progresso e taxa de bytes na importação.
- Ajusta lógica de cálculo do progresso máximo.
- Melhora concorrência na importação com SemaphoreSlim e tarefas.
- Acrescenta variáveis e métodos para controle de tempo e texto de progresso.

+++ module5/Importador/Importador/Logic/ContasLogic.cs
@@ -20,7 +20,7 @@ namespace Importador.Logic
             string[] valores = line.Split(';');
 
             int conta = int.Parse(valores[0]);
-            decimal valor = decimal.Parse(valores[1]);
+            double valor = double.Parse(valores[1]);
             string descricao = valores[2];
 
             // Adiciona os dados ao gridData
